+++ src/Analyzers/CSharp/DiagnosticIdentifiers.Generated.cs
@@ -218,6 +218,7 @@ namespace Roslynator.CSharp
         public const string RemoveEmptySyntax = "RCS1259";
         public const string AddOrRemoveTrailingComma = "RCS1260";
         public const string DisposeResourceAsynchronously = "RCS1261";
+        public const string UnnecessaryRawStringLiteral = "RCS1262";
         public const string InvalidReferenceInDocumentationComment = "RCS1263";
     }
 }